Move proxy_processor into own file
